requirements.txt
poetry
streamlit-chat 
langchain
langchain-openai
langchain-community
sentence-transformers
openai
pypdf
chromadb
tiktoken
pillow
networkx
streamlit-shadcn-ui
pinecone-client
llama-index     
llama-index-embeddings-openai 
llama-index-vector-stores-pinecone
langchain-pinecone
langchainhub
pysqlite3-binary